Update 8rd Gate for Sinar Pure Foods International Go Live Non Fish Modules 5th Commit for Finished Goods Modules
Updates:
+ Purchasing
+ Non-Fish
	+ Transfer Slip
		/ sws_code and ts_code to varchar(50)
+ Fish
+ Finished Goods
	+ Reclassification
		+ Create document function
		+ List function
+ Data-Driven Documents
+ Report

+++ src/java/com/app/wms/engine/db/dao/TsDao.java
@@ -10,7 +10,7 @@ public interface TsDao {
 
     public void insert(Ts t);
 
-    public void updateStockInventory(String productCode, int qty);
+    public void updateStockInventory(String productCode, BigDecimal qty);
 
     public List<Sws> findWhereNotInTs();
